Don't use cloud attributes to discover ips
Cloud attributes don't have a consistent structure among the
different cloud providers. For example:

* In Google Compute engine, cloud['local_ipv4'] and cloud['public_ipv4']
  are lists of ip addresses (or lists with a single element).
* In DigitalOcean they are hashes that contain an 'ipaddress' key
  with the node's address as a value.

Given that there isn't (yet) a consistent structure for the cloud
namespace, this change reverts its use and just reads the
node['ipaddress'] value. It is not the best approach as it might
get the public address, but it is better to have a working one than
a broken one.

A proper fix should include reading the cloud['provider'] value and
use it to properly parse the contents of the cloud['public_ipv4']
object.