Fix bug introduced in #174 (#182)
This fixes a bug that was introduced in #174 which doesn't honor the fact if the
bind parameter is nil don't include any bind line. In our case, we put in the
bind keyword in the parameters parameter so this added two bind statements and
made haproxy unhappy.

It looks like this was a copypasta originally as this was included in the prior
code.