WIP mass fix probably haproxy 5.0 (#259)
* Update config_global.rb

Made the :log property optionally an array to allow multiple logs

* Update haproxy.cfg.erb

Changed the template to support global[log] now being an array.  This allows multiple logs to be specified.

* added sample test for multiple logs

* added multi_log_branch to travis branch config

* added test with multiple logs

* Add option for install only

Needed to be able to have the cookbook handle the installation
and configuration, but did not want to have haproxy managed
by systemd.

Added install_only boolean property to install recipe. This will only handle
the installation and will not register haproxy as a service on the host
system.

* syntax fixes in kitchen.yml

* added new travis matrix instance for config-3-logs

* added an extra level of .each iteration for the global log array

* trying to check for ruby type

* added the extra array iteration

* correcting the array iteration

* try another type

* removed branch from travis build

* corrected the readme and change log

* updating haproxy.service

* updating changelog

* moving to options

* merging haproxy systemd unit from source

* haproxy 1.5 on centos did not support resolvers

* updating changelog

* updating to 1.7.8

* fixing versions

* fixing namespace clashing issue

* fixing changelog

* changing custom_config to source

* custom-template

* adding the directory for chroot